+++ packages/holocron-module-route/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.2](https://github.com/americanexpress/holocron/compare/holocron-module-route@1.10.1...holocron-module-route@1.10.2) (2024-02-22)
+
+**Note:** Version bump only for package holocron-module-route
+
+
+
+
+
+## [1.10.1](https://github.com/americanexpress/holocron/compare/holocron-module-route@1.10.0...holocron-module-route@1.10.1) (2024-02-22)
+
+**Note:** Version bump only for package holocron-module-route
+
+
+
+
+
+# [1.10.0](https://github.com/americanexpress/holocron/compare/holocron-module-route@1.9.2...holocron-module-route@1.10.0) (2024-02-21)
+
+
+### Features
+
+* prepare for react 18 and redux 5 ([#172](https://github.com/americanexpress/holocron/issues/172)) ([19d0755](https://github.com/americanexpress/holocron/commit/19d075578e129e2c1affccc9c3da18e815c3f5fa))
+
+
+
+
+
 ## [1.9.2](https://github.com/americanexpress/holocron/compare/holocron-module-route@1.9.1...holocron-module-route@1.9.2) (2023-11-16)
 
 **Note:** Version bump only for package holocron-module-route